% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataOrdinal}
\alias{dataOrdinal}
\title{Prices of 50,000 round cut diamonds.}
\format{
There are four matrix
\describe{
  \item{dataOrdinal}{min1, max6}
  \item{misspecifiedQmatrix}{min1, max6}
  \item{Qmatrix}{min1, max6}
  \item{respMatrix}{min1, max6}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
dataOrdinal
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
