% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.datalist.R
\name{Generate.datalist}
\alias{Generate.datalist}
\title{Generate and prepare the data list used for this Packgae}
\usage{
Generate.datalist(Qmatrix, response.matrix, GroupID = NA)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{response.matrix}{The response matrix without any missing values. Each cell represents the response of
each item for each person.}

\item{GroupID}{default is NA. Used for multi-group LDCM.}
}
\value{
a list including Y (response matrix), Np (number of observation) and Ni (number of iterm)
}
\description{
\code{Generate.datalist} is used to generate a list for estimation.
}
\examples{
\dontrun{
Qmatrix<-cbind(Qmatrix,rep(1,9));Qmatrix[1,1]<-0
}
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
