% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parm.name.R
\name{Parm.name}
\alias{Parm.name}
\title{Generate Stan code and Run the estimation for LCDM}
\usage{
Parm.name(Qmatrix)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}
}
\value{
a. stan file saved at the specified path
}
\description{
The StanLCDM.script Function to automate Stan code geneartion for LCDMs with binary resposnes
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
