% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanDINA.run.R
\name{StanDINA.run}
\alias{StanDINA.run}
\title{A function to generate leave-one-out cross-validation for Stan Model}
\usage{
StanDINA.run(Qmatrix, response.matrix, script.path = NA,
  savepath = getwd(), savename = "DINA_uninf", iter = 1000,
  warmup = floor(iter/2), chains = 3, init.list = "random",
  control.list = NA)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{savepath}{save the .stan file to somewhere; the default path is getwd()}

\item{savename}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
The StanLCDM.loofit Function to automate Stan code geneartion for LCDMs with binary resposnes
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
