% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanDINO.run.R
\name{StanDINO.run}
\alias{StanDINO.run}
\title{DCM calibration under the DINA model via Stan}
\usage{
StanDINA.run<-function(Qmatrix,response.matrix,script.path=NA,save.path=getwd(),save.name="DINA_uninf",iter=1000,warmup = floor(iter/2),
chain.num=3,init.list='random',control.list=NA)
}
\arguments{
\item{Qmatrix}{A required matrix}

\item{response.matrix}{save the .stan file to somewhere; the default path is getwd()}

\item{script.path}{save the .stan file to somewhere; the default path is getwd()}

\item{save.name}{the name of saved stan file}

\item{iter}{number of iteration of MCMC estimation. defalts to 1000.}

\item{chain.num}{number of MCMC chain.num.}

\item{init.list}{the initial values. 'random' or 'CDM'}

\item{control.list}{the controlled parameters}
}
\value{
StanDINA returens an object of class StanDINA. Methods for StanDINA objects include
\code{\link{extract}} for extract for extracting various components, \code{\link{coef}} for
extracting strctural parameters.
}
\description{
\code{StanDINA} uses Stan program to calibrate the deterministic inputs, noisy and gate model for dichotomous responses, and its
extension

In addition, users are allowed to specify design matrix and link function for each item, and distinct models may be used
in a single test for different items. The attributions can be either dichomous or polytomous.
}
\examples{
\dontrun{
#----------- DINO model-----------#
mod1<-StanDINO.run(Qmatrix, respMatrix, iter=20, init.list='cdm', chain.num = 3)
summary(mod1)
}
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu} \cr Jihong Zhang, University of Iowa, \email{jihong-zhang@uiowa.edu}}
}
