% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanDINO.script.R
\name{StanDINO.script}
\alias{StanDINO.script}
\title{Generate Stan code and Run the estimation for DINO model}
\usage{
StanDINA.run<-function(Qmatrix,response.matrix,script.path=NA,savepath=getwd(),
savename="DINA_uninf",iter=1000,warmup = floor(iter/2),
chains=3,init.list='random',control.list=NA)
}
\arguments{
\item{Qmatrix}{A required matrix}

\item{savename}{name the .stan}

\item{response.matrix}{save the .stan file to somewhere; the default path is getwd()}

\item{script.path}{save the .stan file to somewhere; the default path is getwd()}

\item{iter}{name the .stan}

\item{chains}{name the .stan}

\item{init.list}{name the .stan}

\item{control.list}{name the .stan}
}
\value{
StanDINA returens an object of class StanDINA. Methods for StanDINA objects include
\code{\link{extract}} for extract for extracting various components, \code{\link{coef}} for
extracting strctural parameters.
}
\description{
The \code{StanDINO.run} Function allows to automate Stan code geneartion for DINO model with binary resposnes
}
\examples{
\dontrun{
#----------- DINO model-----------#
mod1<-StanDINO.run(Qmatrix, respMatrix, iter=20, init.list='cdm', chains = 3)
summary(mod1)
}
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu} \cr Jihong Zhang, University of Iowa, \email{jihong-zhang@uiowa.edu}}
}
