% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanLCDM.run.R
\name{StanLCDM.run}
\alias{StanLCDM.run}
\title{Run Stan for estimationg LCDM}
\usage{
StanLCDM.run(Qmatrix, response.matrix, script.path = NA,
  save.path = getwd(), save.name = "LCDM_uninf", iter = 1000,
  warmup = 0, chain.num = 3, init.list = "random",
  control.list = NA)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{response.matrix}{the data matrix.}

\item{script.path}{name the .stan}

\item{save.path}{the directory the Stan Script file save to.}

\item{save.name}{the file name for Stan script}

\item{iter}{number of iteration for MCMC estimation. 1000 by default.}

\item{warmup}{number of iterations for warmup.}

\item{chain.num}{number of MCMC chains}

\item{init.list}{random by default. It could be "cdm" or "random". If "cdm" was chosen,
the starting values will be estimated using maximum-likehood based estimation.}

\item{control.list}{constrains put on the LCDM model.}
}
\value{
MCMC stan object.
}
\description{
The StanLCDM.run function aims to automate runing Stan program to estimate LDCM.
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
