% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanPrior.update.R
\name{StanPrior.update}
\alias{StanPrior.update}
\title{Generate Stan code and Run the estimation for LCDM}
\usage{
StanPrior.update(
  priorUpdate.matrix,
  script.path,
  save.path = getwd(),
  save.name = NA
)
}
\arguments{
\item{priorUpdate.matrix}{the Q-matrix specified for the LCDM}

\item{script.path}{path to save the script file}

\item{save.path}{save the .stan file to somewhere; the default path is getwd()}

\item{save.name}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
The StanPrior.update Function to automate update prior
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
