% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.datalist.R
\name{Generate.datalist}
\alias{Generate.datalist}
\title{Generate Stan code and Run the estimation for LCDM}
\usage{
Generate.datalist(Qmatrix, response.matrix)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{savepath}{save the .stan file to somewhere; the default path is getwd()}

\item{savename}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
The StanLCDM.script Function to automate Stan code geneartion for LCDMs with binary resposnes
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
