% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Install.package.R
\name{Install.package}
\alias{Install.package}
\title{Install and load packages.}
\usage{
Install.package(needed_packages)
}
\arguments{
\item{needed_packages}{the packages to be installed and loaded.}
}
\description{
\code{Install.package} allows other functions to install and load packages.
}
\examples{
\dontrun{
Install.package("ggplot2")
}

}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
