% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanDCM.ppmc.R
\name{StanDCM.ppmc}
\alias{StanDCM.ppmc}
\title{A function to calculate pecentage of extreme o-values for item-pair odds ratios}
\usage{
StanDCM.ppmc(stan.model, response.matrix, n.sim = 6000, n.burnin = 20)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{savepath}{save the .stan file to somewhere; the default path is getwd()}

\item{savename}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
The StanDCM.ppmc Function to automate Stan code geneartion for LCDMs with binary resposnes
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
