% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanDINA.run.R
\name{StanDINA.run}
\alias{StanDINA.run}
\title{DCM calibration under the DINA model via Stan}
\usage{
StanDINA.run(
  Qmatrix,
  response.matrix,
  script.path = NA,
  save.path = getwd(),
  save.name = "DINA_uninf",
  iter = 1000,
  warmup = 0,
  chain.num = 3,
  init.list = "random",
  control.list = NA
)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{response.matrix}{the response matrix}

\item{script.path}{the path to save the .stan file to somewhere; the default path is getwd()}

\item{save.path}{the path to save the .stan file to somewhere; the default path is getwd()}

\item{save.name}{name of the .stan}

\item{iter}{number of iterations}

\item{warmup}{number of warmup iterations}

\item{chain.num}{number of MCMC chains. Default is 3.}

\item{init.list}{initial variables. Default is random. Other options include cdm}

\item{control.list}{constrains put on the multigroup DINO model.}
}
\value{
StanDINA returens an object of class StanDINA. Methods for StanDINA objects include
\code{\link{extract}} for extract for extracting various components, \code{\link{coef}} for
extracting strctural parameters.
}
\description{
\code{StanDINA} uses Stan program to calibrate the deterministic inputs, noisy and gate model for dichotomous responses, and its
extension

In addition, users are allowed to specify design matrix and link function for each item, and distinct models may be used
in a single test for different items. The attributions can be either dichomous or polytomous.
}
\examples{
\dontrun{
#----------- DINA model-----------#
mod1<-StanDINA.run(Qmatrix, respMatrix, iter=20, init.list='cdm', chain.num = 3)
summary(mod1)
}
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu} \cr Jihong Zhang, University of Iowa, \email{jihong-zhang@uiowa.edu}}
}
