% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanPrior.update.R
\name{StanPrior.update}
\alias{StanPrior.update}
\title{Generate Stan code and Run the estimation for LCDM}
\usage{
StanPrior.update(priorUpdate.matrix, script.path, save.path = getwd(),
  save.name = NA)
}
\arguments{
\item{Qmatrix}{the Q-matrix specified for the LCDM}

\item{savepath}{save the .stan file to somewhere; the default path is getwd()}

\item{savename}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
The StanLCDM.script Function to automate Stan code geneartion for LCDMs with binary resposnes
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
